Initial signals implementation
